Fixes XSRF token, moves more HTTP details to api.nw
If the first request is a POST, PUT or DEL request, we need an XSRF
token that we haven't acquired yet (due to not having logged in).

Before weblogin, we always logged in first or when session (and XSRF
token) expired, so we always had an XSRF token. Now we need to make a
GET request to trigger a log in, if we don't have one.

We also move some HTTP details from the LadokSession documentation to
the API documentation.